requirements.txt
requests
aiohttp
asyncio
python-dotenv
tqdm
